Extract save to file logic to save_to_file_module
